30_API/03_exercices/tp_api_babyfoot/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.36
PyJWT==2.9.0
